ha_data_to_mysql/ha_data_mysql/requirements.txt
websockets
mysql-connector-python